added esp8266 source code
added functions to display from the LCD screen in I2C
added NFC reader thanks to a PN532 in UART
NFC reader can read the card CSN.

functions to configure the ESP by using it as an access point.
The configuration page to access is 192.168.4.1 when connected to the AP
When configured, the ESP can read a CSN and will send it to the server page.
Once sent, it will receive the name of the student and display it on the screen